backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
google-cloud-aiplatform==1.38.0
google-generativeai==0.3.0
numpy==1.24.3
opencv-python==4.8.1.78
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.7
pgvector==0.2.3
minio==7.2.0
pydantic==2.5.0
httpx==0.25.2
Pillow==10.1.0
piexif==1.1.3
